Combine DRS and non-DRS atomic scheduler, using cordyceps
